// FE.Core.FW/src/apis/models/BlacklistTopikModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type BlacklistTopikModel = {
  fullName: string;
  dateOfBirth: string;
  identityCard?: string | null;
  citizenIdentityCard?: string | null;
  passport?: string | null;
  type?: number;
  startWorkDate?: string | null;
  finishWorkDate?: string | null;
  notifyResultDate?: string | null;
  finishPunishmentDate?: string | null;
  punishmentAction?: string | null;
  candicateNumber?: string | null;
  examPeriod?: string | null;
  country?: string | null;
  area?: string | null;
  location?: string | null;
  exam?: string | null;
  status?: number;
  note?: string | null;
};